backend/requirements.txt
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
djoser==2.1.0
Pillow==9.2.0
reportlab==3.6.11